requirements.txt
appdirs==1.4.3
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.1.0
django-braces==1.14.0
django-constance==2.8.0
django-filter==2.2.0
django-picklefield==3.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
Markdown==3.2
pathspec==0.7.0
Pillow==7.1.2
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
regex==2020.1.8
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
